model/src/requirements.txt
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
channels==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
constantly==15.1.0
contourpy==1.1.1
crypto==1.4.1
cryptography==41.0.4
cuda-python==12.2.0
cycler==0.11.0
Cython==3.0.2
daphne==3.0.2
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
Django==3.2.5
dnspython==2.4.2
easyocr==1.7.1
executing==1.2.0
filelock==3.13.1
Flask==3.0.0
fonttools==4.42.1
fsspec==2023.10.0
hyperlink==21.0.0
idna==3.4
imageio==2.32.0
imutils==0.5.4
incremental==22.10.0
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mpmath==1.3.0
Naked==0.1.32
nest-asyncio==1.5.8
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pycryptodome==3.19.0
Pygments==2.16.1
pymongo==4.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.3
seaborn==0.12.2
service-identity==23.1.0
shapely==2.0.2
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tifffile==2023.9.26
torch==2.1.1
torchaudio==2.0.2+cu117
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.5
wcwidth==0.2.6
Werkzeug==3.0.1
zope.interface==6.0
